xit1/odevzdani/requirements.txt
pillow==12.1.0
numpy==2.4.1
matplotlib==3.10.8